feat: Implement authentication pages with Suspense and Server Components
